+++ x/vesting/keeper/gov_test.go
@@ -5,54 +5,54 @@ import (
 	testutiltx "github.com/evmos/evmos/v14/testutil/tx"
 )
 
-func (s *KeeperTestSuite) TestGovClawbackStore() {
-	s.SetupTest()
+func (suite *KeeperTestSuite) TestGovClawbackStore() {
+	suite.SetupTest()
 
-	addr := sdk.AccAddress(s.address.Bytes())
+	addr := sdk.AccAddress(suite.address.Bytes())
 
 	// check that the address is not disabled by default
-	disabled := s.app.VestingKeeper.HasGovClawbackDisabled(s.ctx, addr)
-	s.Require().False(disabled, "expected address not to be found in store")
+	disabled := suite.app.VestingKeeper.HasGovClawbackDisabled(suite.ctx, addr)
+	suite.Require().False(disabled, "expected address not to be found in store")
 
 	// disable the address
-	s.app.VestingKeeper.SetGovClawbackDisabled(s.ctx, addr)
+	suite.app.VestingKeeper.SetGovClawbackDisabled(suite.ctx, addr)
 
 	// check that the address is disabled
-	disabled = s.app.VestingKeeper.HasGovClawbackDisabled(s.ctx, addr)
-	s.Require().True(disabled, "expected address to be found in store")
+	disabled = suite.app.VestingKeeper.HasGovClawbackDisabled(suite.ctx, addr)
+	suite.Require().True(disabled, "expected address to be found in store")
 
 	// delete the address
-	s.app.VestingKeeper.DeleteGovClawbackDisabled(s.ctx, addr)
+	suite.app.VestingKeeper.DeleteGovClawbackDisabled(suite.ctx, addr)
 
 	// check that the address is not disabled
-	disabled = s.app.VestingKeeper.HasGovClawbackDisabled(s.ctx, addr)
-	s.Require().False(disabled, "expected address not to be found in store")
+	disabled = suite.app.VestingKeeper.HasGovClawbackDisabled(suite.ctx, addr)
+	suite.Require().False(disabled, "expected address not to be found in store")
 }
 
-func (s *KeeperTestSuite) TestGovClawbackNoOps() {
-	s.SetupTest()
+func (suite *KeeperTestSuite) TestGovClawbackNoOps() {
+	suite.SetupTest()
 
-	addr := sdk.AccAddress(s.address.Bytes())
+	addr := sdk.AccAddress(suite.address.Bytes())
 	addr2 := sdk.AccAddress(testutiltx.GenerateAddress().Bytes())
 
 	// disable the address
-	s.app.VestingKeeper.SetGovClawbackDisabled(s.ctx, addr)
+	suite.app.VestingKeeper.SetGovClawbackDisabled(suite.ctx, addr)
 
 	// a duplicate entry should not panic but no-op
-	s.app.VestingKeeper.SetGovClawbackDisabled(s.ctx, addr)
+	suite.app.VestingKeeper.SetGovClawbackDisabled(suite.ctx, addr)
 
 	// check that the address is disabled
-	disabled := s.app.VestingKeeper.HasGovClawbackDisabled(s.ctx, addr)
-	s.Require().True(disabled, "expected address to be found in store")
+	disabled := suite.app.VestingKeeper.HasGovClawbackDisabled(suite.ctx, addr)
+	suite.Require().True(disabled, "expected address to be found in store")
 
 	// check that address 2 is not disabled
-	disabled = s.app.VestingKeeper.HasGovClawbackDisabled(s.ctx, addr2)
-	s.Require().False(disabled, "expected address not to be found in store")
+	disabled = suite.app.VestingKeeper.HasGovClawbackDisabled(suite.ctx, addr2)
+	suite.Require().False(disabled, "expected address not to be found in store")
 
 	// deleting a non-existent entry should not panic but no-op
-	s.app.VestingKeeper.DeleteGovClawbackDisabled(s.ctx, addr2)
+	suite.app.VestingKeeper.DeleteGovClawbackDisabled(suite.ctx, addr2)
 
 	// check that the address is still not disabled
-	disabled = s.app.VestingKeeper.HasGovClawbackDisabled(s.ctx, addr2)
-	s.Require().False(disabled, "expected address not to be found in store")
+	disabled = suite.app.VestingKeeper.HasGovClawbackDisabled(suite.ctx, addr2)
+	suite.Require().False(disabled, "expected address not to be found in store")
 }